Slight refactoring to move scottplot code to library
