+++ Directory.Packages.Props
@@ -37,6 +37,7 @@
     <PackageVersion Include="OpenAI" Version="2.1.0" />
     <PackageVersion Include="Parquet.Net" Version="5.1.1" />
     <PackageVersion Include="PowerShellStandard.Library" Version="5.1.1" />
+    <PackageVersion Include="ScottPlot" Version="5.0.55" />
     <PackageVersion Include="ScottPlot.WPF" Version="5.0.55" />
     <PackageVersion Include="ShapeCrawler" Version="0.64.3" />
     <PackageVersion Include="Spectre.Console" Version="0.49.1" />
